Remove functionality in observer to put its packets into multiple queues
Formerly, the observer was able to put packets to several receivers.
This commit removes this functionality, as it isn't used at least in any
of the example configs.

A better way to achive this functionality is to create a SplicingConnection
which would push the packet into the seperate queues by itself. This could
then reused in every other module which needs this functionality.