Implement a ReConnector which could reuse the instances from another config
The idea is to first have a normal graph and all modules/instances are
connected. Some time later we want to load a different config, derived
from the original one. In this case we could use the ReConnector, which
checks for all the common nodes (and likewise the common modules) and
transfers the old modules into the freshly created new graph.

This technic allows us to e.g. load a new config (only a filter changed)
to reuse the Observer and all other modules without recreating them from
scratch.

Remark: For now, we disconnect *every* used module, but we reconnect them
right again. This could be furhter optimised, but I think this is enough
for now, as it should be really quick to reconnect them again.

Some sort of sequence diagramm:

1) vermont runs and a new config should be loaded. We parsed the config
   into the 'New Config' graph, but we didn't create any modules yet.

  Old Config:				      | New Config
  Graph:   observer -> queue -> filter1       | observer -> queue -> filter2
  Modules: (ObsInst)->(QInst)-> (Filter1Inst) | (NULL)      (NULL)   (NULL)

2) Disconnect Modules we could reuse in the New Config (topological, so no
   that the queue is first disconnected from filter1 and *then* from the
   observer)

  Graph:   observer -> queue -> filter1       | observer -> queue -> filter2
  Modules: (ObsInst)  (QInst)  (Filter1Inst)  | (NULL)      (NULL)   (NULL)

3) Transfer reusable modules to the new config

  Graph:   observer -> queue -> filter1	      | observer -> queue -> filter2
  Modules: ()          ()      (Filter1Inst)  | (ObsInst)    (QInst)   (NULL)

4) Connect the modules as the same way the config graph is connected and
  create missing module instances.

                                              | observer -> queue -> filter2
                                              | (ObsInst)->(QInst)-> (Filter2Inst)