First shot at implementing a test for reconfiguration
This is more a demonstration for how to link the modules together
(Tobi asked for it). Notice that I actually can't run the test, because
I always get "Test setup error: test tree is empty" as error message if I
try to run it. This could be related with the newer boost library (1.34.1)
I have here and the commit r1000 I needed to make the tests compile.

If someone more familiar with the boost lib and/or using boost version 1.34.1
could have a look at it, I would appreciate it.