added field anonymisationType that enables the individual flagging of fields whether they are anonymized or not (needed for IpfixCsExporter)
small description:
Field anonymisationType always refers to the preceding field and describes whether anonymisation was used, or not. By default, anonymisationType is set to 0 (non-anonymized).
When IpfixRecordAnonymizer anonymises a field, it determines whether the succeding field in the DataRecord is of type anonymisationType. If yes, the field is automatically set to 1.